requirements.txt
numpy
pandas
opencv-python
matplotlib
tensorflow
scikit-learn

torch>=2.0.0
torchvision>=0.15.0
Flask>=2.3.0
Pillow>=9.5.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
tqdm>=4.65.0



requests==2.31.0
Pillow==10.2.0
python-dotenv==1.0.0